+++ src/core/json/json_object.cc
@@ -2,25 +2,23 @@
 // See LICENSE for licensing terms.
 //
 
-#include "core/json_object.h"
+#include "core/json/json_object.h"
 
 #include "base/logging.h"
-#include "core/compact_object.h"
 
 using namespace jsoncons;
 using namespace std;
 
 namespace dfly {
 
-optional<JsonType> JsonFromString(string_view input) {
+optional<JsonType> JsonFromString(string_view input, PMR_NS::memory_resource* mr) {
   error_code ec;
   auto JsonErrorHandler = [](json_errc ec, const ser_context&) {
     VLOG(1) << "Error while decode JSON: " << make_error_code(ec).message();
     return false;
   };
 
-  json_decoder<JsonType> decoder(
-      std::pmr::polymorphic_allocator<char>{CompactObj::memory_resource()});
+  json_decoder<JsonType> decoder(std::pmr::polymorphic_allocator<char>{mr});
   json_parser parser(basic_json_decode_options<char>{}, JsonErrorHandler);
 
   parser.update(input);
