+++ src/server/rdb_load.h
@@ -3,7 +3,6 @@
 //
 #pragma once
 
-#include <jsoncons/json.hpp>
 #include <system_error>
 
 extern "C" {
@@ -13,7 +12,6 @@ extern "C" {
 #include "base/io_buf.h"
 #include "base/mpsc_intrusive_queue.h"
 #include "base/pod_array.h"
-#include "core/json_object.h"
 #include "io/io.h"
 #include "redis/rdb.h"
 #include "server/common.h"
