+++ src/server/search/doc_accessors.cc
@@ -10,7 +10,6 @@
 #include <jsoncons/json.hpp>
 #include <jsoncons_ext/jsonpath/jsonpath.hpp>
 
-#include "core/json_object.h"
 #include "core/search/search.h"
 #include "core/search/vector_utils.h"
 #include "core/string_map.h"
